pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roost-os"
version = "0.1.0"
description = "Revenue Intelligence Agent for short-term rental portfolio"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40",
    "pandas>=2.1",
    "pydantic>=2.5",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["src"]
